refactored both pages
